populate comments from wp_comments
WordPress database referential integrity is preserved reusing its primary keys ids instead of querying for each object:
In WP comments are related to posts which can be of type post, page or attachment (revisions will be ignored for now),
these match cyclope models article and staticpage (attachments will be related contents).
Since cyclope comments rely on the contenttypes framework https://docs.djangoproject.com/en/1.9/ref/contrib/contenttypes/
which relates the object pk to the post or page pk and a contenttype object specyfing which model it is,
and since that info lays in the same wp_post table, each comment would need to query for his post to build the relation.
Instead of that, we create models using the same IDs they had in WP.
And we query only once per content_type for post IDs belonging to it, then we create comments in just as many bulk transactions.